requirements.txt
kafka-python==2.0.2
pyspark==3.5.4
apache-airflow==2.10.4
pytest==8.3.4
fastapi==0.94.0
uvicorn==0.21.0
numpy==1.26.0
pandas==1.5.3
seaborn==0.12.2
matplotlib==3.7.1
scipy==1.10.1
scikit-learn==1.3.2
tensorflow-intel==2.18.0
xgboost==2.1.2
bayesian-optimization==1.4.2

